Ensure travis also runs testvector+namespace tests
This breaks the build, as Kyber and Dilithium currently are not
properly namespaced. More pressingly, this is not a long-term
solution: because of the way Makefile dependencies work,
one failing test now prevents all others from running.

We probably want to change the testing setup not to rely so heavily
on the Makefile, and instead call a Python or shell script.
Related, we may want to reconsider the single scheme / all targets;
this is becoming quite a mess in the Makefile, already.