Enable address and UB sanitizers
This enables address sanitizer and undefined behaviour sanitizer,
but disables leak detection sanitizer.
It seema there is an issue in qemu that we are using, which prevents
spawning leaksanitizer's tracer thread. But as leak detection is
already done by valgrind on x86 build, this functionality is not
absolute need at the moment.