Fix Falcon signature computation
Falcon computes the signature in a loop, retrying if the signature is
too long. The Falcon-1024 limit seems a bit stricter than the Falcon-512
limit, which seemingly hits the recomputation more often.

If it would try to try again for a signature, the implementation would
read from the `hm` buffer again. However, `hm` was specified as a union
with `sig` and the previously attempted signature would be in its place.

In a previous version of Falcon, this recomputation would never happen,
so the case was never hit. Additionally, with the larger signature
lengths present in the NIST reference code and previous PQClean versions
of Falcon, this case would also never be hit.

Thanks a lot to @pornin for finding out this problem.