Add SPDX headers and copyright notices at the beginning of files.
In downstream projects, files may be copied and moved around. This makes
it difficult to track the copyrights and licenses without having them
stored in each file.

I've tried my best to be correct but I had to retrace changes over time,
find origins again and make assumptions.

I've limited myself to Falcon for two reasons. First, at the moment, my
ultimate goal is Debian packaging for liboqs and it is one of the two
relevant algorithms in pqclean for liboqs. Second, it's already quite a
lot of changes and things to potentially discuss.

BTW, sphincsplus in pqclean is also relevant for liboqs packaging but
I've elected to first do changes in sphincs/sphincplus through
sphincs/sphincsplus#64 (even if the two repositories have diverged).

Finally, I'm adding mentions of NIST in copyrights for the reasons I
menetion in the sphincsplus PR: it's what NIST is now using in
https://csrc.nist.gov/Projects/pqc-dig-sig/standardization/example-files
and it's more sensible than the previous mention of a single person.