Replace macros in ML-DSA-87 clean/avx2 API
I apparently forgot to replace the macros in two of the nine `api.h`
files of the ML-DSA implementations in commit
ae29fd8d4f1e00e3786f8b3500306b821a9bd5cd. The previous patch did indeed
fix the issue at hand, which was about missing function symbols, because
the function symbols are indeed declared in `sign.h` and defined in
`sign.c` of each implementation. However, consumers of `api.h` would
still see the old macros for the clean and the avx2 implementations of
ML-DSA-87 and thus still will not be able to construct function pointers
to the relevant functions when including these particular header files.

Refs: https://github.com/PQClean/PQClean/issues/576
Refs: https://github.com/PQClean/PQClean/pull/591